Add require replacement to babel plugin

+++ packages/babel-plugin/src/util/compile.js
@@ -7,20 +7,35 @@ const plugin = require("..");
 const macroPath = path.resolve(__dirname, "..", "..", "client.macro");
 const outputPath = path.join(__dirname, "run-on-server-id-mappings.js");
 
-const transform = (code) => {
+const transform = (code, idMappingsEnabled, evalRequireEnabled) => {
   rimraf.sync(outputPath);
-  const result = babel.transform(code, { plugins: [[plugin, { outputPath }]] });
-  const output = fs.readFileSync(outputPath, "utf-8");
-  rimraf.sync(outputPath);
+  const result = babel.transform(code, {
+    plugins: [
+      [
+        plugin,
+        {
+          idMappings: { enabled: idMappingsEnabled, outputPath },
+          evalRequire: { enabled: evalRequireEnabled },
+        },
+      ],
+    ],
+  });
+  let output = "";
+  if (fs.existsSync(outputPath)) {
+    output = fs.readFileSync(outputPath, "utf-8");
+    rimraf.sync(outputPath);
+  }
   return {
     code: result.code,
     output,
   };
 };
 
-const compile = (code) => {
+const compile = (code, options = {}) => {
+  const { idMappingsEnabled = false, evalRequireEnabled = false } = options;
+
   delete require.cache[macroPath];
-  return transform(code);
+  return transform(code, idMappingsEnabled, evalRequireEnabled);
 };
 
 module.exports = compile;
